crm-agents/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "crm-agents"
version = "0.1.0"
description = "Agentic AI CRM — Python Agent Services (Orchestrator, Memory, Context, Guardrails)"
requires-python = ">=3.12"

dependencies = [
    # Google ADK + Gemini
    "google-adk>=0.3.0",
    "google-cloud-aiplatform>=1.55.0",
    "google-auth>=2.29.0",

    # API Framework
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "sse-starlette>=2.1.0",
    "httpx>=0.27.0",

    # A2A Protocol
    "a2a-sdk>=0.3.0",

    # Data validation
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.1",

    # Vector DB
    "qdrant-client>=1.9.1",

    # Redis
    "redis[hiredis]>=5.0.4",

    # OpenTelemetry
    "opentelemetry-sdk>=1.24.0",
    "opentelemetry-api>=1.24.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.24.0",
    "opentelemetry-instrumentation-fastapi>=0.45b0",
    "opentelemetry-instrumentation-httpx>=0.45b0",

    # PII detection and redaction (Guardrails)
    "presidio-analyzer>=2.2.354",
    "presidio-anonymizer>=2.2.354",
    "spacy>=3.7.4",

    # Structured logging
    "structlog>=24.2.0",

    # Async utilities
    "anyio>=4.3.0",
    "asyncio-throttle>=1.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.6",
    "pytest-cov>=5.0.0",
    "respx>=0.21.1",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "types-redis",
]

[tool.ruff]
line-length = 100
target-version = "py312"
select = ["E", "F", "W", "I", "N", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.mypy]
python_version = "3.12"
strict = false
ignore_missing_imports = true
